separate tloc-list
