+++ catalystwan/endpoints/configuration/policy/list/tloc.py
@@ -5,17 +5,9 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import TLOCList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.tloc import TLOCList, TLOCListEditPayload, TLOCListInfo
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class TLOCListEditPayload(TLOCList, PolicyListId):
-    pass
-
-
-class TLOCListInfo(TLOCList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPolicyTLOCList(APIEndpoints, PolicyListEndpoints):
